prototype: JS-only library on top of core onSafeAreaInsetsChange
When core reports safe area insets through the onSafeAreaInsetsChange
view prop (facebook/react-native#57967), the library needs no native
code:

- SafeAreaProvider renders a plain View with the prop; the payload is
  identical to onInsetsChange.
- SafeAreaView promotes the existing web implementation to every
  platform: insets come from the provider context and are applied
  per-edge as padding or margin, preserving the native shadow node
  semantics (off/additive/maximum). Core's synchronous event dispatch
  keeps both same-frame.

Gated on the RN version so older versions keep the native path. The
jest environment reports RN 1000.0.0, so the test suite exercises the
new path.
